Tests renamed and fixed
