+++ LanguageFeatures/Patterns/type_inference_A20_t02.dart
@@ -15,6 +15,7 @@
 /// - M does not define that operator,
 /// - C is not assignable to the operator's parameter type,
 /// - or if the operator's return type is not assignable to bool.
+///
 /// Else the operator is == or !=. It is a compile-time error if C is not
 /// assignable to T? where T is M's == method parameter type.
 /// The language screens out null before calling the underlying == method, which
@@ -22,8 +23,9 @@
 /// the right, this compile-time error can only happen if a user-defined class
 /// has an override of == with a covariant parameter.
 ///
-/// @description Check that it is a compile-time error if C is not assignable to
-/// T? where T is M's == method parameter type.
+/// @description Check that it is a compile-time error if the operator is a
+/// comparison (<, <=, >, or >=), M does define that operator but C is not
+/// assignable to the operator's parameter type
 /// @author sgrekhov22@gmail.com
 
 // SharedOptions=--enable-experiment=patterns
@@ -31,26 +33,35 @@
 class C {
   final int v;
   const C(this.v);
-
-  @override
-  bool operator ==(covariant C other) => this.v == other.v;
 }
 
+const zero = C(0);
 const one = C(1);
 const two = C(2);
+const three = C(3);
 
 String test1() {
   switch (42) {
-    case == one:
+    case < zero:
 //       ^^^^^^
 // [analyzer] unspecified
 // [cfe] unspecified
-      return "==";
-    case != two:
+      return "<";
+    case <= one:
 //       ^^^^^^
 // [analyzer] unspecified
 // [cfe] unspecified
-      return "!=";
+      return "<=";
+    case > three:
+//       ^^^^^^^
+// [analyzer] unspecified
+// [cfe] unspecified
+      return ">";
+    case >= two:
+//       ^^^^^^
+// [analyzer] unspecified
+// [cfe] unspecified
+      return ">=";
     default:
       return "default";
   }
@@ -58,24 +69,40 @@ String test1() {
 
 String test2() {
   return switch (42) {
-    == one => "==",
+    < zero => "<",
 //  ^^^^^^
 // [analyzer] unspecified
 // [cfe] unspecified
-    != two => "!=",
+    <= one => "<=",
 //  ^^^^^^
 // [analyzer] unspecified
 // [cfe] unspecified
-    _ => "default";
+    > three => ">",
+//  ^^^^^^^
+// [analyzer] unspecified
+// [cfe] unspecified
+    >= two => ">=",
+//  ^^^^^^
+// [analyzer] unspecified
+// [cfe] unspecified
+    _ => "default"
   };
 }
 
 String test3() {
-  if (42 case == one) return "==";
+  if (42 case < zero) return "<";
 //            ^^^^^^
 // [analyzer] unspecified
 // [cfe] unspecified
-  if (42 case != two) return "!=";
+  if (42 case <= one) return "<=";
+//            ^^^^^^
+// [analyzer] unspecified
+// [cfe] unspecified
+  if (42 case > three) return ">";
+//            ^^^^^^^
+// [analyzer] unspecified
+// [cfe] unspecified
+  if (42 case >= two) return ">=";
 //            ^^^^^^
 // [analyzer] unspecified
 // [cfe] unspecified
