+++ LanguageFeatures/Patterns/type_inference_A20_t03.dart
@@ -17,41 +17,55 @@
 /// - or if the operator's return type is not assignable to bool.
 /// Else the operator is == or !=. It is a compile-time error if C is not
 /// assignable to T? where T is M's == method parameter type.
+///
 /// The language screens out null before calling the underlying == method, which
 /// is why T? is the allowed type. Since Object declares == to accept Object on
 /// the right, this compile-time error can only happen if a user-defined class
 /// has an override of == with a covariant parameter.
 ///
 /// @description Check that it is a compile-time error if the operator is a
-/// comparison (<, <=, >, or >=) and M does not define that operator
+/// comparison (<, <=, >, or >=), M does define that operator but it's return
+/// type is not bool
 /// @author sgrekhov22@gmail.com
 
 // SharedOptions=--enable-experiment=patterns
 
+class M {
+  final int v;
+  const M(this.v);
+
+  int operator <(int other) => this.v - other;
+  int operator <=(int other) => this.v - other;
+  int operator >(int other) => this.v - other;
+  int operator >=(int other) => this.v - other;
+}
+
+const zero = M(0);
+const one = M(1);
+const two = M(2);
+const three = M(3);
+
 String test1() {
-  String s = "";
-  switch(s) {
+  switch (zero) {
     case < 0:
 //       ^^^
 // [analyzer] unspecified
 // [cfe] unspecified
       return "<";
-    case <= 0:
+    case <= 1:
 //       ^^^^
 // [analyzer] unspecified
 // [cfe] unspecified
       return "<=";
-    case > 0:
+    case > 3:
 //       ^^^
 // [analyzer] unspecified
 // [cfe] unspecified
-
       return ">";
-    case >= 0:
+    case >= 2:
 //       ^^^^
 // [analyzer] unspecified
 // [cfe] unspecified
-
       return ">=";
     default:
       return "default";
@@ -59,44 +73,42 @@ String test1() {
 }
 
 String test2() {
-  String s = "";
-  return switch(s) {
-    < 0 => "<",
+  return switch (one) {
+    < 1 => "<",
 //  ^^^
 // [analyzer] unspecified
 // [cfe] unspecified
-    <= 0 => "<=",
+    <= 1 => "<=",
 //  ^^^^
 // [analyzer] unspecified
 // [cfe] unspecified
-    > 0 => ">",
+    > 3 => ">",
 //  ^^^
 // [analyzer] unspecified
 // [cfe] unspecified
-    >= 0 => ">=",
+    >= 2 => ">=",
 //  ^^^^
 // [analyzer] unspecified
 // [cfe] unspecified
-    _ => "default";
+    _ => "default"
   };
 }
 
 String test3() {
-  String s = "";
-  if (s case < 0) return "<";
-//           ^^^
+  if (two case < 0) return "<";
+//             ^^^
 // [analyzer] unspecified
 // [cfe] unspecified
-  if (s case <= 0) return "<=";
-//           ^^^^
+  if (one case <= 1) return "<=";
+//             ^^^^
 // [analyzer] unspecified
 // [cfe] unspecified
-  if (s case > 0) return ">";
-//           ^^^
+  if (three case > 0) return ">";
+//               ^^^
 // [analyzer] unspecified
 // [cfe] unspecified
-  if (s case >= 0) return ">=";
-//           ^^^^
+  if (two case >= 2) return ">=";
+//             ^^^^
 // [analyzer] unspecified
 // [cfe] unspecified
   return "default";
