+++ LanguageFeatures/Patterns/type_inference_A20_t04.dart
@@ -15,6 +15,7 @@
 /// - M does not define that operator,
 /// - C is not assignable to the operator's parameter type,
 /// - or if the operator's return type is not assignable to bool.
+///
 /// Else the operator is == or !=. It is a compile-time error if C is not
 /// assignable to T? where T is M's == method parameter type.
 /// The language screens out null before calling the underlying == method, which
@@ -22,92 +23,61 @@
 /// the right, this compile-time error can only happen if a user-defined class
 /// has an override of == with a covariant parameter.
 ///
-/// @description Check that it is a compile-time error if the operator is a
-/// comparison (<, <=, >, or >=), M does not define that operator but it's
-/// return type is not bool
+/// @description Check that it is a compile-time error if C is not assignable to
+/// T? where T is M's == method parameter type.
 /// @author sgrekhov22@gmail.com
 
 // SharedOptions=--enable-experiment=patterns
 
-class C {
+class M {
   final int v;
-  const C(this.v);
+  const M(this.v);
 
-  int operator <(int other) => this.v - other;
-  int operator <=(int other) => this.v - other;
-  int operator >(int other) => this.v - other;
-  int operator >=(int other) => this.v - other;
+  @override
+  bool operator ==(covariant M other) => this.v == other.v;
 }
 
-const zero = C(0);
-const one = C(1);
-const two = C(2);
-const three = C(3);
+const one = M(1);
+const two = M(2);
 
 String test1() {
-  switch (42) {
-    case < zero:
-//       ^^^^^^
+  switch (one) {
+    case == 1:
+//       ^^^^
 // [analyzer] unspecified
 // [cfe] unspecified
-      return "<";
-    case <= one:
-//       ^^^^^^
+      return "==";
+    case != 2:
+//       ^^^^
 // [analyzer] unspecified
 // [cfe] unspecified
-      return "<=";
-    case > three:
-//       ^^^^^^^
-// [analyzer] unspecified
-// [cfe] unspecified
-      return ">";
-    case >= two:
-//       ^^^^^^
-// [analyzer] unspecified
-// [cfe] unspecified
-      return ">=";
+      return "!=";
     default:
       return "default";
   }
 }
 
 String test2() {
-  return switch (42) {
-    < zero => "<",
-//  ^^^^^^
+  return switch (two) {
+    == 1 => "==",
+//  ^^^^
 // [analyzer] unspecified
 // [cfe] unspecified
-    <= one => "<=",
-//  ^^^^^^
+    != 2 => "!=",
+//  ^^^^
 // [analyzer] unspecified
 // [cfe] unspecified
-    > three => ">",
-//  ^^^^^^^
-// [analyzer] unspecified
-// [cfe] unspecified
-    >= two => ">=",
-//  ^^^^^^
-// [analyzer] unspecified
-// [cfe] unspecified
-    _ => "default";
+    _ => "default"
   };
 }
 
 String test3() {
-  if (42 case < zero) return "<";
-//            ^^^^^^
+  if (one case == 1) return "==";
+//             ^^^^
 // [analyzer] unspecified
 // [cfe] unspecified
-  if (42 case <= one) return "<=";
-//            ^^^^^^
-// [analyzer] unspecified
-// [cfe] unspecified
-  if (42 case > three) return ">";
-//            ^^^^^^^
-// [analyzer] unspecified
-// [cfe] unspecified
-  if (42 case >= two) return ">=";
-//            ^^^^^^
+  if (two case != 2) return "!=";
+//             ^^^^
 // [analyzer] unspecified
 // [cfe] unspecified
   return "default";
