+++ LanguageFeatures/Patterns/type_inference_A22_t01.dart
