+++ LanguageFeatures/Patterns/type_inference_A25_t01.dart
